+++ packages/components/src/components/select/select/select.tsx
@@ -49,6 +49,7 @@ import type { InputSearchInputEventDetail } from '../../input-search/input-searc
 import { getComponentCss } from './select-styles';
 import {
   getSelectedOptionString,
+  type SelectChangeEventDetail,
   type SelectDropdownDirection,
   type SelectOptgroup,
   type SelectOption,
@@ -136,7 +137,15 @@ export class Select {
   /** The id of a form element the select should be associated with. */
   @Prop({ reflect: true }) public form?: string; // The ElementInternals API automatically detects the form attribute
 
+  /** Emitted when the select has lost focus. */
+  @Event({ bubbles: false }) public blur: EventEmitter<void>;
+
   /** Emitted when the selection is changed. */
+  @Event({ bubbles: true }) public change: EventEmitter<SelectChangeEventDetail>;
+
+  /**
+   * @deprecated since v3.30.0, will be removed with next major release, use `change` event instead. Emitted when the selection is changed.
+   */
   @Event({ bubbles: false }) public update: EventEmitter<SelectUpdateEventDetail>;
 
   @State() private isOpen = false;
@@ -302,6 +311,7 @@ export class Select {
           disabled={this.disabled}
           onClick={this.onComboClick}
           onKeyDown={this.onComboKeyDown}
+          onBlur={this.onComboBlur}
           ref={(el) => (this.buttonElement = el)}
         >
           {this.slottedImagePath && <img src={this.slottedImagePath} alt="" />}
@@ -319,6 +329,7 @@ export class Select {
           popover="manual"
           tabIndex={-1}
           onToggle={() => this.onToggle()}
+          onBlur={(e: any) => e.stopPropagation()}
           role="dialog"
           aria-label={this.label}
           aria-hidden={this.isOpen ? null : 'true'}
@@ -337,6 +348,8 @@ export class Select {
               theme={this.theme}
               onInput={this.onFilterInput}
               onKeyDown={this.onComboKeyDown}
+              onBlur={(e: any) => e.stopPropagation()}
+              onChange={(e: any) => e.stopPropagation()}
               ref={(el: HTMLPInputSearchElement) => (this.inputSearchElement = el)}
             />
           )}
@@ -383,6 +396,7 @@ export class Select {
   private onClickOutside = (e: MouseEvent): void => {
     if (this.isOpen && isClickOutside(e, this.buttonElement) && isClickOutside(e, this.popoverElement)) {
       this.isOpen = false;
+      this.blur.emit();
     }
   };
 
@@ -537,6 +551,10 @@ export class Select {
   };
 
   private emitUpdateEvent = (): void => {
+    this.change.emit({
+      value: this.value,
+      name: this.name,
+    });
     this.update.emit({
       value: this.value,
       name: this.name,
@@ -553,6 +571,7 @@ export class Select {
   };
 
   private onFilterInput = (e: CustomEvent<InputSearchInputEventDetail>): void => {
+    e.stopPropagation();
     const { hasFilterResults, resetCurrentlyHighlightedOption } = updateFilterResults(
       this.selectOptions,
       this.selectOptgroups,
@@ -570,6 +589,14 @@ export class Select {
           this.inputSearchElement.focus();
         });
       });
+    }
+  };
+
+  private onComboBlur = (e: FocusEvent): void => {
+    e.stopPropagation();
+    // Don't emit blur when opening the dropdown
+    if (!this.isOpen) {
+      this.blur.emit();
     }
   };
 }